chore: a big cleanup and bud fixing in all of the files

+++ src/services/api/ours/favoriteMemoriesService.ts
@@ -1,7 +1,14 @@
 import axios from "axios";
 import { BASE_URL } from "../../../configuration/config";
 
-export async function getFavoriteMemoryById(token: string, memoryId: string) {
+export async function getFavoriteMemoryById(
+  token: string | null,
+  memoryId: string
+) {
+  if (!token) {
+    return;
+  }
+
   const res = await axios.get(
     `${BASE_URL}/favorite-memories/get-favorite-memory/${memoryId}`,
     {
@@ -14,7 +21,11 @@ export async function getFavoriteMemoryById(token: string, memoryId: string) {
   return res.data.memory;
 }
 
-export async function getAllFavoriteMemories(token: string) {
+export async function getAllFavoriteMemories(token: string | null) {
+  if (!token) {
+    return;
+  }
+
   const res = await axios.get(
     `${BASE_URL}/favorite-memories/get-favorite-memories`,
     {
@@ -27,7 +38,11 @@ export async function getAllFavoriteMemories(token: string) {
   return res.data.memories;
 }
 
-export async function getRecentFavoriteMemories(token: string) {
+export async function getRecentFavoriteMemories(token: string | null) {
+  if (!token) {
+    return;
+  }
+
   const res = await axios.get(`${BASE_URL}/favorite-memories/recent`, {
     headers: {
       Authorization: `Bearer ${token}`,
@@ -38,10 +53,14 @@ export async function getRecentFavoriteMemories(token: string) {
 }
 
 export async function createFavoriteMemory(
-  token: string,
+  token: string | null,
   memory: string,
   date: string
 ) {
+  if (!token) {
+    return;
+  }
+
   const res = await axios.post(
     `${BASE_URL}/favorite-memories/add-favorite-memory`,
     { memory, date },
@@ -56,11 +75,15 @@ export async function createFavoriteMemory(
 }
 
 export async function updateFavoriteMemory(
-  token: string,
+  token: string | null,
   memoryId: string,
   memory: string,
   date: string
 ) {
+  if (!token) {
+    return;
+  }
+
   const res = await axios.put(
     `${BASE_URL}/favorite-memories/edit-favorite-memory/${memoryId}`,
     { memory, date },
@@ -74,7 +97,14 @@ export async function updateFavoriteMemory(
   return res.data.memory;
 }
 
-export async function deleteFavoriteMemory(token: string, memoryId: string) {
+export async function deleteFavoriteMemory(
+  token: string | null,
+  memoryId: string
+) {
+  if (!token) {
+    return;
+  }
+
   await axios.delete(
     `${BASE_URL}/favorite-memories/delete-favorite-memory/${memoryId}`,
     {
